src/http_iot_streamer/requirements.txt
Flask==3.0.0
google-cloud-monitoring==2.18.0
black==23.12.1
